fix: refactor tasks structure
