+++ psycop/common/sequence_models/tasks/__init__.py
@@ -1,0 +1,9 @@
+from .task_registry import create_behrt, create_encoder_for_clf
+from .tasks import BEHRTForMaskedLM, EncoderForClassification
+
+__all__ = [
+    "BEHRTForMaskedLM",
+    "EncoderForClassification",
+    "create_behrt",
+    "create_encoder_for_clf",
+]
